backend/pyproject.toml
# ./backend/pyproject.toml

[tool.poetry]
name = "web-notepad-backend"
version = "0.1.0"
description = ""
authors = ["Орлов Ярослав <orlov.avis@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
asyncpg = "^0.25.0"
bcrypt = "^3.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
alembic = "^1.8.0"
psycopg2-binary = "^2.9.3"
databases = "^0.6.0"
aiopg = "^1.3.4"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
orm = {extras = ["postgresql"], version = "^0.3.1"}

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = {version = "^22.3.0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
